Separate xcm config file (#1317)
* Separate XCM config to separate file: moonbase

* Xcm config to a separate file: moonriver

* Remove ParentAsSuperUser in moonbase as its not needed

* Xcm config to separate file: Moonbeam

* Remove parentassuperuser from imports

* Moonbase: fmt + fix tests

* Moonriver:fix tests

* Fix moonbeam tests

* Fix conflicts

* Remove unused dep

* Clippy changes

* Fix unclosed delimiter

* Fix moonbase xcm config

* remove useless tests

* Add correct test glmr
